+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - fix: loading images fit screen width
 - fix: cookie manager, when going back in history
 - fix: buttons not reachable on some dialogs
+- fix: Auth-request dialog
 
 
 ### v 9.5